MegaPortraits/requirements.txt
torch
torchvision
Pillow
numpy
opencv-python
matplotlib
pyyaml